Add hack to lock flow of scroll events to window where scroll started
A bit evil but there is probably no other way, because this is very
different from GDK's usual behavior.

There's one quirk left, if you start a scroll and move to another GTK+
window in the same process and click, the click won't be processed until
the folow of momentum events has ended.