Introduce a background window
The background window covers the part left uncovered by the overshoot
window. Two background windows are used to be able to cover an "L shape"
that appears when is scrolled in the horizontal and vertical direction
at the same time. These background windows are used to capture events,
such as scroll events. In the future, it could also be used to draw a
specific background pattern/gradient (if the window is configured as
INPUT/OUTPUT window).