Make scrolled window work well with Mac touchpad
Modify the scrolled window code to work with a Mac touchpad as you
would expect. When precise deltas are received from the Mac touchpad,
overshooting and snapping back will work. The existing kinetic scrolling
code is only used for snapping back. The amount of pixels to overshoot
is determined by the incoming scroll events from the touchpad. We use
the deceleration mechanism that was already in place to snap back.

Other changes made are:
 - Increased the kinetic scrolling distance, this makes things feel
 nicer IMHO.
 - Removed everything related to handling kinetic scrolling by making
 drag gestures. We don't need this code.