Improve overshooting behavior
This is done by taking the stiffness calculations as found in the WebKit
source code. These calculations are now used to compute the overshooting
distance as well as handling the snap back animation. Also, we only
start overshooting after we have hit a border in the container (i.e. an
adjustment has reached its upper/lower bound).

Also rename the timeout functions from deceleration to snap back, since
these are not handling kinetic deceleration in this code.