Quartz: Make event loop deal with recursive poll invocations
This patch addresses a number of problems with the recursion in
poll_func() that happens on OS X. Firstly, autorelease pool refreshes
are done from prepare and only if we are at the base level of *both*
main loops and not recursing in poll. This first part incorporates
suggestions from Daniel Sabo and John Ralls.

Secondly, a mechanism is implemented to detect if poll has been
called recursively. This is used to discard file descriptor state
at the base level, retrieved before possibly recursing (through the
NSApp get next event).