Defer map for GtkNSView
This is an attempt to solve an issue in GtkNSView that occurs in
Xamarin Studio when we try to use an animating NSProgressIndicator.
The error only occurred if the indicator was in animating mode,
and it seems likely to be a threading race condition somewhere.

This patch does two things. I've found the issue to be reproducible
if we do one of these things but not the other, but I have not been
able to reproduce it with both of these:
  1. Slightly defer the mapping of the widget
  2. Lock the view during clip_to_parent_viewports()

https://bugzilla.xamarin.com/show_bug.cgi?id=17401