dnd: Take widget selection code from the Quartz backend
The current code does a simple check on the hierarchy of the
first widget found to be under the pointer, whereas the quartz
code does a more thorough check that gets things right if
multiple widgets lie on the same coordinates.

This fixes Drag n' drop (specially the latter) on the MonoDevelop
autohide docks.

https://bugzilla.xamarin.com/show_bug.cgi?id=11281